[aws-lambda] Some docs, split up types and tests (#42080)
* [aws-lambda] Some docs, split up index.d.ts into multiple files.

Documents handler type and conventions for adding triggers more
explicitly. The same index.d.ts docs should explain the thinking for the
new layout convention.

No changes to the type tests in this first commit, to show that the tests
still passs, and because I think they deserve more attention (and I need
lunch right now).

Most event structures have little to no changes other than re-ordering or removing
inaccurately copy/pasted comments in a case or two.

Loses history, but github doesn't support rename history anyway, and
it's even more ugly to get git to track file splits even for local
usage. Apologies for the huge diff.

* [aws-lambda] Split tests to per-service files, add contributor README.md

Updates most of the files to a consistent style.
