+++ types/aws-lambda/trigger/cloudwatch-logs.d.ts
@@ -1,0 +1,33 @@
+import { Handler } from "../hander";
+
+export type CloudWatchLogsHandler = Handler<CloudWatchLogsEvent, void>;
+
+/**
+ * See http://docs.aws.amazon.com/lambda/latest/dg/eventsources.html#eventsources-cloudwatch-logs
+ */
+export interface CloudWatchLogsEvent {
+    awslogs: CloudWatchLogsEventData;
+}
+
+export interface CloudWatchLogsEventData {
+    data: string;
+}
+
+export interface CloudWatchLogsDecodedData {
+    owner: string;
+    logGroup: string;
+    logStream: string;
+    subscriptionFilters: string[];
+    messageType: string;
+    logEvents: CloudWatchLogsLogEvent[];
+}
+
+/**
+ * See http://docs.aws.amazon.com/AmazonCloudWatch/latest/logs/SubscriptionFilters.html#LambdaFunctionExample
+ */
+export interface CloudWatchLogsLogEvent {
+    id: string;
+    timestamp: number;
+    message: string;
+    extractedFields?: { [key: string]: string };
+}
