+++ types/aws-lambda/tsconfig.json
@@ -18,6 +18,19 @@
     },
     "files": [
         "index.d.ts",
-        "aws-lambda-tests.ts"
+        "aws-lambda-tests.ts",
+        "test/alb-tests.ts",
+        "test/api-gateway-tests.ts",
+        "test/cloudformation-tests.ts",
+        "test/cloudfront-tests.ts",
+        "test/cloudwatch-tests.ts",
+        "test/codepipeline-tests.ts",
+        "test/cognito-tests.ts",
+        "test/dynamodb-tests.ts",
+        "test/kinesis-tests.ts",
+        "test/lex-tests.ts",
+        "test/s3-tests.ts",
+        "test/sns-tests.ts",
+        "test/sqs-tests.ts"
     ]
 }
